vm_tools: set --strict-balloon on sibling VMs
Enable strict ballooning for sibling VMs, which disables deflate-on-oom
in the sibling VMs. This is necessary because manatee memory service
needs to be able to precisely manage VM memory, and deflate-on-oom
bypasses that control.

Disabling deflate-on-oom also fixes the oom killer in sibling VMs. Since
deflate-on-oom results in the balloon slightly deflating whenever the
guest is about to oom, it prevents all oom killing when the balloon is
not empty. Since the balloon is unlikely to completely deflate in
manatee, deflate-on-oom effectively prevents all oom killing, live
locking the guest under memory pressure.

BUG=b:233970160
TEST=tast run brya-manatee multivm.Lifecycle.crostini

Change-Id: I19618371dae26a4af2f60e731ae5465f93abd253
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3682863
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>