shill: remove route and neighbor connection diagnostics
This patch removes from ConnectionDiagnostics all the routines
pertaining to routes and to neighbor reachability:

Routes: ConnectionDiagnostics only checks the existence of a route for a
given destination but does not check the validatity of the route itself.
Any issue with a missing route in the DUT's routing configuration of a
particular network is already detected in ConnectionDiagnostics in the
ICMP or DNS routines in the form of ENETUNREACH errors or similar, as
well as in PortalDetector which always precedes ConnectionDiagnostics.
Therefore these checks are redundant and can be removed for
simplification and maintenance.

Neighbor reachability: patchpanel's link monitor has now been enabled on
all physical networks where neighbor reachability is possible. It reacts
dynamically to changes in reachability of relevant neighbors (gateway
and DNS) while ConnectionDiagnostics only do static lookups in the
neighbor table which can return stale data. These routines can also be
removed without loss of information.

IP collisions: the Network.Shill.ConnectionDiagnosticsIssue UMA metric
shows there is exactly 0 event recorded for this case and it never
happens. ARP timeouts for the gateway are already covered by the link
monitor.

Removing these routines from ConnectionDiagnostics allows to keep only DNS
and ICMP based routines and simplify the logic. Only keeping routines
above layer 3 allows to consistently use ConnectionDiagnostics for any
kind of network connection including Cellular or VPN connections.

A future patch will change the scheduling logic in ConnectionDiagnostics
such that:
  - all DNS and gateway addresses are pinged over both IPv4 and IPv6,
  - all DNS servers are used for resolving some test hostnames over both
  IPv4 and IPv6,
to achieve better exhaustiveness in detecting connectivity issues above
layer 3.

BUG=b:229309479
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I956830631ca171120643ab592606d55c96aeae13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4074985
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>