vm_tools: sommelier: Add XDG output client code
Retrieve the XDG logical coordinates from the host compositor if
this interface is supported. This will be necessary for optimizations
that are intended to resolve some of the issues in the mentioned bug.

BUG=b:217976094
TEST=Ensured that Borealis functioned as before, ensure XDG coords
show up in Wayland debug logs

Change-Id: I8da82d8ab9d70104d0b23c899d9f15be84ee03ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3690005
Reviewed-by: Nic Hollingum <hollingum@google.com>
Auto-Submit: Isaac Bosompem <mrisaacb@google.com>
Tested-by: Isaac Bosompem <mrisaacb@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>