cryptohome: Use loop backing device on hibernate resume
If there's a snapshot set up for encstateful, use that with a plain loop
device backing when resuming from hibernate. This is done because
hiberman sets up and maintains the dm-snapshots early in boot, and to
get a writable device, cryptohome needs to use the snapshot for
encstateful. This replaces the earlier temporary mechanism, which
applied a dm-snapshot underneath LVM itself.

BUG=b:233103309
TEST=Hibernate and resume

Change-Id: Icbd63578ce2490bf576e93e9bc3a2f4d7b434d88
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3700621
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>