patchpanel: ndproxy: treat NS src IP as guest confirmation
NDProxy relies on Neighbor Advertisement message from guest as
confirmation of IPv6 address configured on that guest, adding a /128
route when detecting that. This patch extends it to treat Neighbor
Solicitation message with a global or unique local source IP as a valid
guest IP confirmation as well.

BUG=b:134351098
BUG=b:234298157
TEST=unit

Change-Id: I3bd2330812866cebd0b727c0e2b0f7a18874fae5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3691081
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>