resourced: Check charger status to determine if it's powering the system
Power tests compute power usage by measuring how much the battery
discharged during a test. This requires the test to disable the charger.
The `online` attribute will still be 1 indicating the charger is
still connected, but the `status` attribute will indicate the charger
is no longer charging.

This change makes it so the power tests actually test the governor
changes.

BUG=b:221091499
TEST=ran power_VideoCall.FDO_fast and verified powersave bias was
applied

Change-Id: I7c22b9c38cc7a8044a7a8e8d23842db1b623e5fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3687088
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Vovo Yang <vovoy@chromium.org>