debugd: Make sure cpuidle states are restored when respawning
The perf_tool now uses a temp file instead of a map to keep track
of all cpuidle states during collecting ETM data on ARM. This way
it can properly restore the states if debugd crashes in the middle
of a collection.

The temp file is removed after restoring the states either after a
collection or when debugd initializes perf_tool in case of a crash.
It is also validated to only allow cpuidle states file patterns
before used.

The log_tool has been modified to use the PerfTool instance from
debugd instead of constructing one locally to avoid unnecessary
calls to RestoreCpuIdleStates.

BUG=b:233106663
TEST=tast run $DUT debugd.PerfTool.etm

Change-Id: Ib15c0e00120711684ddf237e06e4b14b27942d7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3766021
Commit-Queue: Tuo Shan <shantuo@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Tuo Shan <shantuo@google.com>