shill: Ethernet ignores empty permanent MAC address
This patch makes Ethernet class ignoring to treat "00:00:00:00:00:00" as
a valid permanent address, since it is not supposed to appear on a real
device.

The change will be helpful for the veth usage in the tast tests.
Currently we use the permanent MAC address as the storage identifier in
the profile, and use the non-permanent one as the fallback. Since a veth
always has an all-zero address as its permanent MAC address (and there
seems no way to change that), if there are two veths on DUT, they will
share the same entry in the profile. After this change, we can let them
use different entries by simply setting different MAC address on them.

BUG=b:192436642
TEST=verified that two veth services no longer share the same entry in
  profile

Change-Id: Ib33c5d47bbbaee83233f2295010d38b3ee178ba6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3695698
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>