cryptohome: Use user snapshot device if it exists
To support hibernate on enterprise enrolled devices Chrome needs
to peek at the hibernate feature flags, which are stored on
cryptohome. As a result the cryptohome data on disk disk is altered,
which causes corruptions in the resuming kernel.

Hiberman already sets up device mapper snapshots for other volumes
that are susceptible to be altered during the resume flow, which
allows to discard the changes if resume succeeds. A hiberman change
adds cryptohome to the LVs for which to create snapshots. In
cryptohome use the snapshot device if it exists, instead of the
original underlying volume.

BUG=b:293895107
TEST=no file system corruptions when resuming from hibernate on
  brya with CL:4767276 and CL:3700622

Change-Id: Icf3d945e3011fbb32930ae58af3503721138ca7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3700622
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>