hiberman: Use correct block device for DiskFile
DiskFile has an assumption that all files opened will be on the
unencrypted stateful partition. With the use of LVM, and specifically a
logical volume for hibernate state, that's no longer true. Fix the
code that finds the block device for the given file to always open
the block device at the hibernate directory, rather than the unencrypted
stateful partition.

BUG=b:232806877
TEST=Hibernate and resume

Change-Id: I4d6f0f61477d4ac0034aa1d4e7c6b2e01af30648
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3699366
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>