hiberman: Move hibernate state directory
The resume-init command needs to mount and access the hibernate logical
volume since that's where the snapshot files live. These files are used
to capture writes to normally writable logical volumes via dm-snapshot.

One of the LVs we install a snapshot on top of is unencrypted stateful.
It presents a problem if the hibernate path of
/mnt/stateful_partition/unencrypted/hibernate is inside one of the
volumes we need to install a snapshot on before mounting. So create a
top level /mnt/hibernate directory where we mount this instead. This
directory exists in the rootfs, so it needs to be created at hiberman
ebuild time. So there's a sister CL in chromiumos-overlay required to
make this change work. Since this directory is created at build time,
also remove the helper function that created the directory.

BUG=b:232806877
TEST=hibernate and resume

Cq-Depend: chromium:3700725
Change-Id: I896d6846422c024c41f5c4b2c437e34a52884a5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3699367
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>