hiberman: Activate hibervol LV if needed
The check in setup_hibernate_lv() was simply checking if the path to the
logical volume exists, and assuming the hibernate volume didn't exist at
all and needed to be created if not. However, that logic fails to take
into account the (very common) case that the hibernate volume exists, it
simply isn't activated at the moment.

Activate the logical volume. Augment the ActivatedLogicalVolume
structure to optionally not deactivate the volume on drop(), since we
keep the hibervol mount after the activated volume struct disappears.

BUG=b:233103309
TEST=hibernate and resume

Change-Id: Ibde4d5dbedd5d1c4a39e636493478a5b24f11f4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3700914
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>