power: Separate out some hibernate metrics
Certain metrics, like SuspendAttemptsBeforeSuccess,
SuspendAttemptsBeforeCancel, and BatteryDischargeRateWhileSuspended
should not be combined between suspend-to-RAM and hibernate. We instead
want to be able to examine those metrics separately for hibernation.
Send a boolean through the result processing indicating whether or not
we hibernated, and use that in the metrics collector to change the name
of those metrics upon resume from hibernation.

BUG=b:232184089
TEST=FEATURES=test emerge-$BOARD power_manager

Change-Id: Iad1e73d77dabc0c28b95fe258b9de1a481c1f271
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3705884
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>