sirenia: Migrate from sys_util to crosvm-base.
This picks up the newer SharedMemory implementation which takes the
intended size in the constructor and handles preallocation.

poll.rs was ported from sys_util since it was dropped in
crosvm-base.

The Facility and Priority enums were included in
libsirenia::linux::syslog instead of depending on crosvm-base since it
declared one of the enums private.

The log crate is now used over sys_util::syslog. Syslog initialization
was moved to a common location so that both stderr and syslog can be
handled appropriately. Additional dependencies were needed to cover the
change.

BUG=b:229016539, b:233694042
TEST=cargo test

Cq-Depend: chromium:3707486
Change-Id: Ied172350b7c5f751c4849b90be6584567c955c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3705887
Reviewed-by: Paramjit Oberoi <psoberoi@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>