shill: Add new Network class
This patch adds a new Network class which currently is a wrapper of the
current Connection class.

After this patch, the Network class will only be an internal class of
the Device class, which means classes other than Device are unaware of
it. This patch cuts the dependency between Device and Connection: all
the calls to Connection from Device will go through the Network class,
but other classes (not Device) will still use the Connection class.

This patch is not supposed to change any logic, the unit tests are not
modified and supposed to be passed as before.

BUG=b:232177767
BUG=b:182777518
TEST=unit tests

Change-Id: I7c99999aa5f4a64a4d27f0ac10befa69b80dc8ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3559096
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>