shill: wifi: Remove connection check in OnNeighborReachabilityEvent()
This patch remove the call to Network::HasConnectionObject() in the WiFi
class when processing a neighbor reachability event. Since we have the
ipconfig check below, this check should be unnecessary. This patch may
affect the logic since Connection may not have the same lifetime with
IPConfig, but by design, ipconfig should have meaningful values iif
layer 3 is ready (not exact considering that ipconfig may be set before
routing layer is configured by shill), so this change should be
expected.

The main purpose of this change is to remove the Connection reference in
wifi_test.cc

BUG=b:232177767
BUG=b:182777518
TEST=unit_tests

Change-Id: I96697e2eac580d23723888cce039f9b2d1942623
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3713480
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>