shill: Initialize Network class with ifindex and ifname
This patch make the Network class storing the ifindex and ifname, since
they are constants for a Network object. This patch only changes the
interaction between Device and Network, and thus only Device unit test
is affected.

This patch is not supposed to change any logic.

BUG=b:232177767
BUG=b:182777518
TEST=unit tests

Change-Id: I0fc69450b83967005c15744979d0927a582d84be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3714149
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>