mojo_service_manager: Log loaded policy files
Log which policy files are loaded. It is set to VLOG(1) so can be used
for debugging.
Move the policy checking after creating `Daemon` so the environment will
be created when running it. Without this, Missing AtExitManager error
will happen if log level is set to -1.

BUG=b:236798109
TEST=emerge-betty-pi-arc mojo_service_manager
TEST=(On DUT) mojo_service_manager --check_policy --log_level=-1

Change-Id: I6e8256289a0b1c470d132054751eec1ffeaaaa2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3717714
Commit-Queue: Chung-Sheng Wu <chungsheng@google.com>
Tested-by: Chung-Sheng Wu <chungsheng@google.com>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>