dlp: Cleanup database on startup.
On older Linux kernels (before 5.1) there is no way
to be notified about files deletion through fanotify.

Because of that we need to construct another way to remove
non-existed anymore inodes from the database that is done
on the daemon startup.

On recent kernels it won't be critical as entries will be
purged once the file is deleted.

BUG=chromium:1169864
TEST=cros_workon_make --board=$BOARD chromeos-base/dlp --test

Change-Id: I3f34c6c0f8eb2831826d8c81fc09c58313d1381a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3716585
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>