shill: docs: Add white background to images
The documentation images have a transparent background. This does not
work very well when using dark mode on a browser, since the background
becomes black and it conflicts with the dark colors in the drawing.
Forcing a white background will ensure that the contents of the image
are always visible.
To avoid having big images on git, create the new image files using
GraphViz and save the exported SVG file into GS bucket.

BUG=None
TEST=Visually check the images on VSCode

Change-Id: I4f5180cc980ca53333af44d7a5e983af40344c85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3718808
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>