hiberman: Don't complete the Resume dbus call early
When Chrome makes a ResumeFromHibernate() dbus call to hiberman,
hiberman doesn't complete the call right away, but returns a
PendingResumeCall structure to the main resume path. This represents a
dbus method that has not been completed, and will get completed
automatically when the PendingResumeCall is dropped. This is the
serialization mechanism where hiberman can stall the rest of boot while
the resume completes.

In our attempt to zero out the hibernate secret after it was used, we
accidentally dropped the pending resume call too early. To fix this,
take a mutable reference rather than taking ownership in the if let
statement that zeroizes the sensitive data.

BUG=b:215415258
TEST=Resume from hibernate, observe boot not continue a bit

Change-Id: I7812526fac0a3b4efdb20875ede854f33b3bb022
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3722075
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Evan Green <evgreen@chromium.org>
Auto-Submit: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>