shill: portal: directly set Service state when skipping detection
If any of the preconditions for starting portal detection are not met,
the Service can be set to the ONLINE state without any considerations
for any other side effect of running portal detection (example:
metrics, starting ConnectivityDiagnostics, ...).

This allows to call SetServiceState() directly from
StartPortalDetection() now that all preconditions are guaranteed to
always be checked in StartPortalDetection() regardless of the original
trigger for portal detection.

The main benefit is to reduce to two the call sites of
SetServiceConnectedState(), inside PortalDetectorCallback().

BUG=b:230339227
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I8b0863c3c9ad4e014ea358ceef4aa8e69d29ce6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3722138
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>