shill: portal: move Service checks post portal detection
Now that SetServiceConnectedState is only called from within
PortalDetectorCallback, the preconditions checks on |selected_service_|
can be moved outside of SetServiceConnectedState to
PortalDetectorCallback. Doing so avoids repeated existence checks on
|selected_service_|.

To be consistent with the preconditions checked in StartPortalDetector,
a check on Network::HasConnectionObject is also added.

BUG=b:230339227
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Ia55af58d2e5f22c212cbbe8e8bd11053eaf30c41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3722139
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>