shill: portal: inline SetServiceConnectedState (2/2)
This patch inlines the remaining logic of SetServiceConnectedState()
inside PortalDetectorCallback() for the case of a failed network
validation where |selected_service_| should transition to the
non-kStateOnline states.

This removes the last call site of SetServiceConnectedState(),
therefore this function can also be removed.

This patch also fixes a bug where OnNetworkValidationFailure() was
illogically called after the next OnNetworkValidationStart(). Note that
this bug was only happening because it is currently not straightforward
to call OnNetworkValidationStart() from Device at the time corresponding
to the added portal detection delay.

Checked on a dual stack network that portal detection starts and retries
correctly for:
  - initial connection
  - dbus-send --system --fixed --print-reply
      --dest=org.chromium.flimflam /
      org.chromium.flimflam.Manager.RecheckPortal
  - dbus-send --system --fixed --print-reply
      --dest=org.chromium.flimflam
      /service/$sevice_id org.chromium.flimflam.Service.SetProperties
      string:"CheckPortal" string:"true"

Checked retry behavior with:
  ip{,6}tables -v -I INPUT -i $iface -p tcp --sport 443 -j DROP

BUG=b:230339227
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I72188f166b0d872fd47621496073cea4b1385755
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3722141
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>