shill: Introduce NetworkConfig class
This patch introduces the NetworkConfig class which will be used to
represent a network configuration. This can represent the config from a
single SOURCE (e.g., static ip config) and can be merged. Differences
with IPConfig are:
- It is able to represent IPv4 and IPv6 configs in a single object;
- All fields in it are optional and thus can represent that some fields
  are not configured.

Currently this class only contains the fields will be used in
StaticIPParameters.

BUG=b:232177767
TEST=build

Change-Id: Ie09edb405f8139c4ad69ddda26e89c408a4707b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3720279
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>