shill: StaticIPParameters uses NetworkConfig
Currently StaticIPParameters uses KeyValueStore to represent the IP
config. This patch changes it to use the strong-typed NetworkConfig
for that purpose.

Note that this patch assumes that in StaticIPParameters, address and
prefix len should always come together (and thus this address is
meaningful). This patch affects the behavior if this is not the case.

Since we now use NetworkConfig to represent the previous IPConfig
before applying static IPConfig, it also contains the |default_routing|
config, and thus b/184533440 will be solved.

This patch is the required to detach IPConfig and StaticIPParameters, so
that we can use the stateless NetworkConfig to represent the
StaticIPParameters.

BUG=b:232177767
BUG=b:184533440
TEST=unit_tests

Change-Id: I199031c44872ec019b08027a7d3a6985229cebd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3720280
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>