hiberman: Pass pending resume object to launch resume
The PendingResumeCall delays the return from the ResumeFromHibernate()
dbus method, which is what prevents the rest of the normal boot flow
from continuing while a hibernate resume is in progress. It's important
that this object stay alive for the full duration of the resume attempt,
otherwise boot will continue concurrently with resume. We've already had
one bug where we accidentally dropped this object too early.

Pass this object into launch_resume_image(), which represents the final
step in the resume process. It's slightly strange since this function
doesn't actually use this object, it simply receives it to enforce that
it stays alive. The goal here is to basically internalize external
constraints, hopefully allowing the compiler to tell us when this object
does not live long enough.

BUG=b:215415258
TEST=cargo build

Change-Id: Ie264590458b905b0618c75136532e3e2a728dfc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3721082
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>