diagnostics: Wait until debugd ready to check support of nvme-self-test
We used to check whether nvme-self-test is available during the
initialization of CrosHealthdRoutineService. The availablility at that
time is not accurate since debugd might have not been ready. We should
wait until debugd to be ready before we check the nvme-self-test.

Also, change the |SystemConfig::NvmeSelfTestSupported| function to be
async so that we can replace the sync D-Bus call by an async one.

BUG=b:186084361, b:216252312, b:228125442
TEST=(cr) USE="-cros-debug" FEATURES=test emerge-volteer diagnostics
TEST=1. Prepare a DUT that prints "DBus.Error.ServiceUnknown" for debugd
        after reboot.
     2. Deploy the new diagnostics package.
     3. Verify the error message is no longer printed.
TEST=1. Prepare a DUT that cannot run nvme_self_test after reboot but
        can run it after cros_healthd is restarted.
     2. Deploy the new diagnostics package.
     3. Verify the nvme_self_test can be run after reboot.

Change-Id: I91139f6a435493f70954a98c5f6225b4500320da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3721786
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Tested-by: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Chung-Sheng Wu <chungsheng@google.com>