sepolicy: Add contexts for /dev/mapper
systemd-tmpfiles reads SELinux contexts from file contexts and becomes
unhappy if we don't have labels for /dev/mapper/*. Add a context for
the directory: in order to preserve behavior, we keep it as
block_device, since that is the default labal that gets assigned to
this directory and contained device nodes. Eventually, we will move
this to a separate domain in order to have more fine-grained
permissions.

BUG=b:235313784
TEST=systemd-tmpfiles --create --boot --clean --prefix /dev doesn't
     show errors.

Change-Id: Ic0f1cca493a142dcc7e9ef07f272f7c59740d0e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3727620
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>