shill: Introduce DevicePortalDetectorTest
The intention here is to introduce a set of simpler Device
unit tests that are less reliant on Mocks, focused on the
PortalDetector state machine.

The existing tests rely heavily on mocks and only test
part of the state machine.

In order to properly test upcoming UMA metrics changes, it
will be necessary to test the larger state machine. The new
test implementations will enable that and should make the
tests easier to maintain.

BUG=b:237545672
TEST=unit tests: DevicePortalDetectorTest.*

Change-Id: I2bc4f907aea68a32a21da28df1b2c17a0757e3dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3732970
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>