cryptohome: Implement AuthBlock::PrepareForRemoval()
Implement AuthBlock::PrepareForRemoval() and call it from AuthBlockUtility.

This abstraction allows auth blocks to perform custom steps before
an old AuthBlockState can be deleted. The only auth block that
implements this is CryptohomeRecoveryAuthBlock: it talks to
PinWeaver to revoke the old secrets (to make it impossible to
decrypt the old AuthBlockState even if the deletion of the old
auth block state file fails).

BUG=b:208357933
TEST=cros_run_unit_tests --board=${BOARD} --packages cryptohome

Change-Id: I7332ae07cfa608d9f313e3d5358405edebb12917
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3714369
Commit-Queue: Anastasiia N <anastasiian@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Anastasiia N <anastasiian@chromium.org>