sepolicy: Rename "minijailed" domain.
The "minijailed" domain is the domain for processes launched from
Minijail running an unlabeled executable. Because we want processes
to run in their own domain (and not all in the "minijailed" domain),
the security.SELinuxProcesses test will fail if it encounters processes
in the "minijailed" domain.

This is problematic because "minijailed" has a positive connotation
(the process is "jailed"), so the test failures are confusing. Address
this by changing the name of the domain to "unconfined_proc", which
better reflects the fact that such processes are unconfined from an
SELinux perspective.

Not only are these processes lacking an SELinux policy, just being
launched by Minijail does not actually sandbox a process. A Minijail
cmdline like:

minijail0 /path/to/unconfined_exec

will result in an unsandboxed process. This CL makes these processes
run in an SELinux domain that reflects this: an unlabeled file launched
by Minijail still ends up in an unconfined domain.

BUG=b:237445387
TEST=Run binary labelled as unconfined in minijail
TEST=(minijail0 /path/to/binary).
TEST=Check domain for binary.

Cq-Depend: chromium:3736284
Change-Id: Ib956653f4ec63dcb19320f60981985c2f771148e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3733433
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>