vm_tools: Set the root cursor in sommelier
The window manager is required to set the cursor for the root window,
otherwise applications that fail to set their own cursor will have an
invisible one.

Crostini has a workaround of calling `xsetroot`, but it is preferable to
do this in C code in sommelier.

It is safe to do this and retain crostini's workaround (see
b/204724669#comment6), since overriding the cursor will be invisible to
users. Going forward we can remove that workaround.

BUG=b:204724669
TEST=sommelier -X glxgears

Change-Id: I7e9d0008461b4211a8dfa15622c22dc75bbd1711
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3734467
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Lucy Qu <lqu@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>