camera: Fix sequence checker DCHECK crash in CameraMonitor
Currently sequence checker DCHECK can be triggered on
CameraMonitor::~CameraMonitor if Attach/Detach is called but
StartMonitorOnThread is never called in between (This can happen when
OpenDevice succeed but ConfigureStream failed).

In this case, the sequence checker is bound to the CameraMonitor.thread_
on SetTaskRunnerOnThread, and never detached, and
base::TimerBase::~TimerBase checks for sequence checker which will fail.

If StartMonitorOnThread is called in between (when camera is working
normally), on CameraMonitor.thread_.Stop() the task will be destroyed
(TimerBase::OnTaskDestroyed), which detach the sequence checker, so the
CameraMonitor::~CameraMonitor works correctly. (Note that
TimerBase::OnTaskDestroyed is removed in upstream libchrome)

Fix this by calling base::OneShotTimer::AbandonAndStop() instead of
base::OneShotTimer::Stop() in CameraMonitor::Detach(), which guarantees
to abandon the tasks and detach the sequence checker. (Note that
currently the DelayTimerBase::Stop() only abandon tasks when some global
feature is enabled, and there's comment about Stop() in the future will
also abandon tasks.)

BUG=b:234085429
TEST=Test camera service with WIP fake camera HAL that errors on OpenDevice.

Change-Id: I40f7a7a0840890c076d9d0dcac8ee7336445468b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3733940
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Auto-Submit: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>