hiberman: Add auth_session ResumeFromHibernate() method
Add a ResumeFromHibernateAS() method that takes in an auth_session
token, which can be used by cryptohome to generate the hibernate secret
on the fly during an auth session. Chrome will call this in between the
authenticate and mount operations so that we can resume before the
user's cryptohome gets mounted.

BUG=b:233103309
TEST=hibernate and resume

Change-Id: I1d97546c4fdbb364b0d5fcd755bea0af7f5320e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3735897
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>