shill: Refactor Service-Network interactions (part 3)
This patch replaces the SetIPConfig() method in Service with a new
SetAttachedNetwork() method. Different with SetIPConfig() which will be
called when the IPConfig object used to setup the connection changed,
the new SetAttachedNetwork() will be called together with
SelectService() in Device, which means that Service will hold a weak ptr
to the Network whose owner (Device) selects this service.

The main usages of the previous SetIPConfig() are:
- Make Service aware of the IPConfig object it needs to expose on the
  D-Bus interface. This part is implemented by reading the corresponding
  field from Network now.
- Set up the callback from Service to notify a change in
  StaticIPParameters. This part is implemented by that Service calls the
  corresponding function directly on Network now.

A new callback is added from Network to Service, which is used to notify
Service of IPConfig changes.

BUG=b:232177767
TEST=unit tests

Change-Id: I81bfab78b621cc090113dc35b84e2caf7555ef02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3736416
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>