vm_tools: sommelier: Add direct scale mode (again)
This is a new mode within Sommelier in which the output
resolution reported to clients will have a simple and direct
relationship with the physical output resolution.

This mode can be enabled with the addition of "--direct-scale"
parameter to the command line.

The scale factor used for this calculation will be used directly,
without factoring in device/user scale settings from the host.

The current scaling method, depending on the user scale setting,
can result in very large display buffers being allocated. This is
not an ideal scenario from a performance and memory footprint standpoint.

BUG=b:217976094
TEST=Verified original/direct scale modes within Borealis and the
original mode under Crostini

Change-Id: I5634c9ac586de01a2d7ef1e187f6559c2c7812f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3739261
Commit-Queue: Isaac Bosompem <mrisaacb@google.com>
Reviewed-by: Chloe Pelling <cpelling@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Tested-by: Isaac Bosompem <mrisaacb@google.com>