vm_tools: Add SetTimezone API in maitred to update VM timezone
The API supports two ways to update the timezone:
- Update symlink
- Bind-mount into /etc/localtime
The first method is general Linux way. The second way is required for
applications that do not work with chained symlinks (which is required
if /etc is mounted RO).

Additional work is required down the line to accept POSIX TZ string
along with the timezone name, so that we can account for zoneinfo file
being out-of-date or missing in VM (see TODO in code).

BUG=b:170959970
TEST=run unit test, launch Borealis, Steam and PoE

Change-Id: I54cf1f1125b44659c81c52d559023da37e772962
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3742560
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Tested-by: Max Lee <endlesspring@google.com>
Commit-Queue: Max Lee <endlesspring@google.com>